Added More Tests
- fixed the null check thanks again @mtuska
- added mutli proc tests as well. still not seeing the issues via TDD.
